Create createDetermenisticRandom function
